downloads

+++ auto-cleanup-home.sh
@@ -9,8 +9,17 @@ HOME="/home/${USER}/"
 echo "Deleting all personal files (for public computers)"
 #
 
+MOUNTCHECK=$(df -h |grep Downloads | wc -l) 
+if test $FSTABCHECK = "1" 
+then
+    echo "Skipping Downloads because of mounted device"
+else
+    rm -r $HOME/Downloads/*
+fi
+ 
+ 
+ 
 
-rm -r $HOME/Downloads/*
 rm $HOME/*
 rm $HOME/.local/share/RecentDocuments/*
 rm $HOME/.kde/share/apps/RecentDocuments/*
